image-service/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Pillow==8.2.0
flask-cors
pika
numpy
requests
prometheus-client
psutil